Changed location of video buffer from 0x8000 to 0x4000. Changed default RAM size from 32k to 64k. Changed timer interrupt code to react off of cpu clocks rather than real time. Extended timer settup ports to port 8 (low) and port 9 (high) to allow for more granular control of timer interrupts. Changes bus_attach routine from BYTE pages to int pages to allow for full access of 256 pages. Extended work on 6502 plugin.

+++ dynosaur/bus.h
@@ -4,7 +4,7 @@
 
 BYTE ram_accessor(WORD address, bool read, BYTE value);
 BYTE bus_accessor(WORD address, bool read, BYTE value);
-void attach_bus(accessor_t accessor, BYTE start_page, BYTE num_pages);
+void attach_bus(accessor_t accessor, BYTE start_page, int num_pages);
 BYTE port_accessor(BYTE port_no, bool read, BYTE value);
 void attach_port(port_accessor_t accessor, BYTE port_no);
 void init_bus();
